lambdas/makeslide/requirements.txt
boto3
google-api-python-client
google-auth
google-auth-httplib2